# Removes non-static locals collection